# docs/swagger.yaml
basePath: /
definitions:
  http.companyResponse:
    properties:
      amount_of_employees:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      registered:
        type: boolean
      type:
        type: string
      updated_at:
        type: string
    type: object
  http.createCompanyRequest:
    properties:
      amount_of_employees:
        type: integer
      description:
        type: string
      name:
        type: string
      registered:
        type: boolean
      type:
        type: string
    type: object
  http.errorResponse:
    properties:
      error:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
    type: object
  http.patchCompanyRequest:
    properties:
      amount_of_employees:
        type: integer
      description:
        type: string
      name:
        type: string
      registered:
        type: boolean
      type:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for managing companies.
  title: Company Service API
  version: "1.0"
paths:
  /companies:
    post:
      consumes:
      - application/json
      description: Creates a company and publishes a company.created event.
      parameters:
      - description: Company payload
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/http.createCompanyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.companyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/http.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.errorResponse'
      security:
      - BearerAuth: []
      summary: Create a company
      tags:
      - companies
  /companies/{id}:
    delete:
      description: Deletes an existing company and publishes a company.deleted event.
      parameters:
      - description: Company ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete a company
      tags:
      - companies
    get:
      description: Returns a company by ID. This endpoint is public.
      parameters:
      - description: Company ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.companyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.errorResponse'
      summary: Get a company
      tags:
      - companies
    patch:
      consumes:
      - application/json
      description: Partially updates an existing company and publishes a company.updated
        event.
      parameters:
      - description: Company ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/http.patchCompanyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.companyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/http.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.errorResponse'
      security:
      - BearerAuth: []
      summary: Patch a company
      tags:
      - companies
  /healthz:
    get:
      description: Returns a simple status response when the process is running.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - health
  /readyz:
    get:
      description: Checks whether the service can reach its database dependency.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/http.errorResponse'
      summary: Readiness check
      tags:
      - health
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and a JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"